third_party/blink/web_tests/http/tests/xmlhttprequest/post-blob-content-type-async-expected.txt

CONSOLE ERROR: Access to XMLHttpRequest at 'http://localhost:8000/xmlhttprequest/resources/access-control-allow-lists.php' from origin 'http://127.0.0.1:8000' has been blocked by CORS policy: No 'Access-Control-Allow-Origin' header is present on the requested resource.
Test verifies that content MIME type is set correctly when Blob is sent using XMLHttpRequest asynchronously.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS postedMimeType is "text/plain;charset=utf-8"
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType is "multipart/mixed;boundary=\"--blob-boundary\""
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS Cross-origin request without CORS headers should fail.
PASS postedMimeType is "text/plain;charset=utf-8"
PASS postedMimeType is "text/plain;charset=utf-8"
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS postedMimeType is "multipart/mixed;boundary=\"--blob-boundary\""
PASS postedMimeType === undefined is true
PASS postedMimeType === undefined is true
PASS successfullyParsed is true

TEST COMPLETE
